Fleshed out the domain model a little
Details
-------
Renamed 'Guru' as 'User', added skeletal 'Sector' and 'Skill' classes.
Added fields to the user domain object, viz. homeTown, aboutMe and
many to many relationships with the skill and sector classes.

Note that the user class is expected to evolve and these changes
have been made to allow development on the front end to progress
which in turn will allow the domain model to evolve some more.

It's a symbiotic relationship.

Ticket
------